+++ src/main/java/ru/olegcherednik/zip4jvm/io/zstd/bit/BitOutputStream.java
@@ -29,14 +29,14 @@ public class BitOutputStream {
             0x3FFFFFFF, 0x7FFFFFFF }; // up to 31 bits
 
     private final byte[] outputBase;
-    private final long outputAddress;
-    private final long outputLimit;
+    private final int outputAddress;
+    private final int outputLimit;
 
     private long container;
     private int bitCount;
     private long currentAddress;
 
-    public BitOutputStream(byte[] outputBase, long outputAddress, int outputSize) {
+    public BitOutputStream(byte[] outputBase, int outputAddress, int outputSize) {
         checkArgument(outputSize >= SIZE_OF_LONG, "Output buffer too small");
 
         this.outputBase = outputBase;
